requirements.txt
langgraph>=0.2.55
langchain>=0.3.14
langchain-openai>=0.2.8
langchain-community>=0.3.12
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
duckduckgo-search>=6.3.0
pydantic>=2.9.0
